extracted updateInventoryCards
